+++ lib/AllocateDescriptorsPass.cpp
@@ -26,9 +26,9 @@
 #include "llvm/Support/CommandLine.h"
 #include "llvm/Support/raw_ostream.h"
 
+#include "Passes.h"
 #include "clspv/AddressSpace.h"
 #include "clspv/Option.h"
-#include "clspv/Passes.h"
 
 #include "ArgKind.h"
 #include "Constants.h"
