+++ lib/ClusterConstants.cpp
@@ -29,9 +29,9 @@
 #include "llvm/Pass.h"
 #include "llvm/Support/raw_ostream.h"
 
+#include "Passes.h"
 #include "clspv/AddressSpace.h"
 #include "clspv/Option.h"
-#include "clspv/Passes.h"
 
 #include "ArgKind.h"
