+++ lib/DirectResourceAccessPass.cpp
@@ -30,8 +30,8 @@
 #include "llvm/Pass.h"
 #include "llvm/Support/raw_ostream.h"
 
+#include "Passes.h"
 #include "clspv/Option.h"
-#include "clspv/Passes.h"
 
 #include "ArgKind.h"
 #include "Constants.h"
