+++ lib/HideConstantLoadsPass.cpp
@@ -23,8 +23,8 @@
 #include "llvm/Pass.h"
 #include "llvm/Support/raw_ostream.h"
 
+#include "Passes.h"
 #include "clspv/AddressSpace.h"
-#include "clspv/Passes.h"
 
 using namespace llvm;
 using std::string;
