+++ lib/InlineFuncWithPointerBitCastArgPass.cpp
@@ -21,7 +21,7 @@
 #include "llvm/Support/raw_ostream.h"
 #include "llvm/Transforms/Utils/Cloning.h"
 
-#include "clspv/Passes.h"
+#include "Passes.h"
 
 using namespace llvm;
