+++ lib/InlineFuncWithPointerToFunctionArgPass.cpp
@@ -19,8 +19,8 @@
 #include "llvm/Support/raw_ostream.h"
 #include "llvm/Transforms/Utils/Cloning.h"
 
+#include "Passes.h"
 #include "clspv/AddressSpace.h"
-#include "clspv/Passes.h"
 
 using namespace llvm;
