+++ lib/InlineFuncWithSingleCallSitePass.cpp
@@ -20,8 +20,8 @@
 #include "llvm/Transforms/Utils/Cloning.h"
 
 #include "ArgKind.h"
+#include "Passes.h"
 #include "clspv/Option.h"
-#include "clspv/Passes.h"
 
 using namespace llvm;
