+++ lib/ReplacePointerBitcastPass.cpp
@@ -19,7 +19,7 @@
 #include "llvm/Pass.h"
 #include "llvm/Support/raw_ostream.h"
 
-#include "clspv/Passes.h"
+#include "Passes.h"
 
 using namespace llvm;
